add and modify
add DiskUsage 
modified disk_collertor
